+++ Vulkan/Source/CommandList.cpp
@@ -290,11 +290,11 @@ namespace RHI::Vulkan
         {
             auto buffer = m_context->m_bufferOwner.Get(command.indexBuffers);
             vkCmdBindIndexBuffer(m_commandBuffer, buffer->handle, 0, VK_INDEX_TYPE_UINT32);
-            vkCmdDrawIndexed(m_commandBuffer, parameters.elementCount, parameters.instanceCount, parameters.firstElement, parameters.vertexOffset, parameters.firstInstance);
+            vkCmdDrawIndexed(m_commandBuffer, parameters.elementCount, parameters.instanceCount, parameters.firstElement, int32_t(parameters.vertexOffset), uint32_t(parameters.firstInstance));
         }
         else
         {
-            vkCmdDraw(m_commandBuffer, parameters.elementCount, parameters.instanceCount, parameters.firstElement, parameters.firstInstance);
+            vkCmdDraw(m_commandBuffer, parameters.elementCount, parameters.instanceCount, parameters.firstElement, uint32_t(parameters.firstInstance));
         }
     }
